Fixed the case when cursor sometimes does not diappear.
Cursor does not disappear when some classes are not applied before the script runs. They apply later and override the 'none' style for the player cursor. Now I use computed cursor style, rather than `element.style.cursor` value. I also added the `primevideo` to domain to the regexp to allow the script to run on that domain. Logging added.